requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Vector Database
qdrant-client==1.16.2

# Microsoft Graph API
msal==1.26.0
httpx==0.26.0
aiohttp==3.9.1

# Document Processing
pypdf2==3.0.1
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.1.4
beautifulsoup4==4.12.2
markdown==3.5.1

# OCR (optional)
pytesseract==0.3.10
pillow==10.2.0

# AI/ML
openai==1.9.0
anthropic==0.18.0
tiktoken==0.5.2

# Text Processing
langchain-text-splitters==0.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==41.0.7
tenacity==8.2.3

# Background Tasks
celery==5.3.4
rq==1.15.1

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Authentication
PyJWT==2.8.0
email-validator==2.1.0
